Fixes to authentication module (#14) (#15)
* Fix matching IPv4-mapped IPv6 to subnet mask (#13)
* Only pass the last IP address from X-Forwarded-For

This function used to stop at the first X-Forwarded-For header line and pass
its value as-is along to authentication modules.  This way of parsing left out
scenarios with multiple header lines with the same name, when it fact they
should have been concatenated to a list and handled from left to right.

This patch changes the behavior so that the function now continues parsing
header lines after the first match and discards all IP addresses apart from the
last one.

Motivation:
  https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/X-Forwarded-For#parsing

---------

Co-authored-by: Ronja Koistinen <ronja.koistinen@helsinki.fi>